requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
boto3==1.28.15
botocore==1.31.15
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coverage==7.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.2.1
django-embed-video==1.4.8
django-environ==0.10.0
django-extensions==3.2.1
django-filter==22.1
django-markdown==0.8.4
django-nine==0.2.7
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.9.8
django-storages==1.13.2
django-taggit==3.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
gunicorn==21.2.0
idna==3.4
importlib-metadata==5.2.0
jmespath==1.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==23.0
pathspec==0.10.1
Pillow==9.4.0
polib==1.1.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-gettext==4.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.13
wcwidth==0.1.9
Werkzeug==2.3.4
zipp==3.11.0
